Additions related to introduction of pawn class
Added a decent amount, moved things around added the Position
class, the Color class, a few board methods. This was all to
allow for the pawn class to generate moves, thusly it needed
to know it's color, whether or not spaces were occupied and
of course where it could move.