ConsoleChessGame preliminary implementation
Game renderers and you can move pieces and stuff.
-Added a new PieceGetter view class that when updated
parses the input from the player and sets the players move
-That move is then grabbed by ConsoleChessGame (new class)
that then selects the piece to be moved, the board is then
rendered. The player classes are also new. Player is basically
an interface that the game queries to get a move. So in a human
players case it would update a view to get its move, and in
a computer players case it would call a simulator to get its
move. Also added a few misc. tests.